requirement.txt
altair==4.0.1
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
beautifulsoup4==4.8.1
bleach==3.1.0
blinker==1.4
boto3==1.11.10
botocore==1.14.10
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
packaging==19.2
pandas==1.0.0
pandocfilters==1.4.2
parso==0.6.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.0
pydeck==0.2.1
Pygments==2.5.2
pyobjc-core==6.1
pyobjc-framework-Cocoa==6.1
pyobjc-framework-FSEvents==6.1
pyparsing==2.4.5
pyrsistent==0.15.7
pysqlite3==0.4.1
pytest==5.3.5
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.1
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.2
scipy==1.4.1
Send2Trash==1.5.0
simpy==3.0.11
six==1.13.0
soupsieve==1.9.5
streamlit==0.54.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.8
validators==0.14.2
watchdog==0.10.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
